configure: use -Werror to make the include GL and GLES2 header check guaranteed fatal
Clang will only give a warning for the redefinition of typedef GLenum.
Since master is build with -Werror this will result in a build failure
later in the gl plugin. Add -Werror to the test, so the test result is as
expected. This will allow the gl plugins to build.

https://bugzilla.gnome.org//show_bug.cgi?id=731692